move loginActivity file according to  mvp
