+++ app/src/main/AndroidManifest.xml
@@ -32,7 +32,7 @@ limitations under the License.
             android:label="@string/app_name"
             android:theme="@style/Theme.Taverna">
         <activity
-                android:name=".activities.LoginActivity"
+                android:name=".ui.login.LoginActivity"
                 android:label="@string/app_name">
         </activity>
         <activity
